requirements.txt
Django==2.2.6
django-webpack-loader==0.6.0
djangorestframework==3.10.3
djangorestframework-jwt == 1.11.0
pytz==2019.3
sqlparse==0.3.0
